[core] Add core headers to private CMake target FILE_SETS
This is a non-functional change, it only improves the IDE experience.

In the future, we could avoid copying headers to build-directory
and verbose target_include_directory statements by just using the file_sets.
Same with INSTALL, would go through headers.
This could ultimately go into the ROOT custom macros, but for the moment
let's only do the minimal changes for the IDE and let the rest for later.